+++ test/subclass-test.js
@@ -2,15 +2,18 @@ var _ = require('lodash');
 var expect = require('./spec-helper').expect;
 var minim = require('../lib/minim');
 
+var ArrayElement = minim.getElementClass('array');
+var StringElement = minim.getElementClass('string');
+
 describe('Minim subclasses', function() {
   // TODO: Provide better interface for extending elements
   var MyElement = function() {
-    minim.StringElement.apply(this, arguments);
+    StringElement.apply(this, arguments);
     this.element = 'myElement';
     this._attributeElementKeys = ['headers'];
   }
 
-  MyElement.prototype = _.create(minim.StringElement.prototype, {
+  MyElement.prototype = _.create(StringElement.prototype, {
     ownMethod: function() {
       return 'It works!';
     }
@@ -54,7 +57,7 @@ describe('Minim subclasses', function() {
     });
 
     it('should create headers element instance', function() {
-      expect(myElement.attributes.get('headers')).to.be.instanceof(minim.ArrayElement);
+      expect(myElement.attributes.get('headers')).to.be.instanceof(ArrayElement);
     });
 
     it('should leave foo alone', function() {
@@ -64,7 +67,7 @@ describe('Minim subclasses', function() {
 
   describe('serializing attributes', function() {
     var myElement = new MyElement();
-    myElement.attributes.set('headers', new minim.ArrayElement(['application/json']));
+    myElement.attributes.set('headers', new ArrayElement(['application/json']));
     myElement.attributes.get('headers').content[0].meta.set('name', 'Content-Type');
 
     it('should serialize headers element', function() {